Change : migration to the original ffmpeg's branch(take the ffmpeg from LAV - as it contains many helpful corrections); thanks to LAV for this "decode H.264 extradata" code; remove many custom ffdshow code; CDDAReader - remove C4018 warning;
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4184 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/reader/CDDAReader/CDDAReader.cpp
@@ -459,7 +459,7 @@ HRESULT CCDDAStream::Read(PBYTE pbBuffer, DWORD dwBytesToRead, BOOL bAlign, LPDW
 					 &BytesReturned, 0);
 		UNUSED_ALWAYS(b);
 
-		size_t l = (size_t)min(min(len, RAW_SECTOR_SIZE - offset), m_llLength - pos);
+		size_t l = (size_t)min(min(len, size_t(RAW_SECTOR_SIZE - offset)), size_t(m_llLength - pos));
 		memcpy(pbBuffer, &buff[offset], l);
 
 		pbBuffer += l;
